Roles and Permissions for Company

+++ routes/api.php
@@ -3,12 +3,7 @@
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\CompanyController;
-use App\Http\Controllers\ProductController;
-use App\Http\Controllers\TeamController;
-use App\Http\Controllers\FloorController;
-use App\Http\Controllers\ProcessController;
-use App\Http\Controllers\ScheduleController;
-use App\Http\Controllers\UserController;
+use App\Http\Controllers\LoginController; 
 
 Route::post('/login', [LoginController::class, 'login']);
 
@@ -16,42 +11,9 @@ Route::get('/user', function (Request $request) {
     return $request->user();
 })->middleware('auth:sanctum');
 
-    Route::get('/company', [CompanyController::class, 'index']);
-    Route::post('/company', [CompanyController::class, 'store']);
-    Route::put('/company/{id}', [CompanyController::class, 'update']);
-    Route::delete('/company/{id}', [CompanyController::class, 'destroy']);
-
-    Route::get('/products', [ProductController::class, 'index']);
-    Route::post('/products', [ProductController::class, 'store']);
-    Route::put('/products/{id}', [ProductController::class, 'update']);
-    Route::delete('/products/{id}', [ProductController::class, 'destroy']);
-
-    Route::get('/teams', [TeamController::class, 'index']);
-    Route::post('/teams', [TeamController::class, 'store']);
-    Route::put('/teams/{id}', [TeamController::class, 'update']);
-    Route::delete('/teams/{id}', [TeamController::class, 'destroy']);
-
-    Route::get('/floors', [FloorController::class, 'index']);
-    Route::post('/floors', [FloorController::class, 'store']);
-    Route::put('/floors/{id}', [FloorController::class, 'update']);
-    Route::delete('/floors/{id}', [FloorController::class, 'destroy']);
-
-    Route::get('/processes', [ProcessController::class, 'index']);
-    Route::post('/processes', [ProcessController::class, 'store']);
-    Route::put('/process/{id}', [ProcessController::class, 'update']);
-    Route::delete('/process/{id}', [ProcessController::class, 'destroy']);
-
-    Route::get('/schedules', [ScheduleController::class, 'index']);
-    Route::post('/schedules', [ScheduleController::class, 'store']);
-    Route::put('/schedules/{id}', [ScheduleController::class, 'update']);
-    Route::delete('/schedules/{id}', [ScheduleController::class, 'destroy']);
-
-    Route::get('/resources', [ResourceController::class, 'index']);
-    Route::post('/resources', [ResourceController::class, 'store']);
-    Route::put('/resources/{id}', [ResourceController::class, 'update']);
-    Route::delete('/resources/{id}', [ResourceController::class, 'destroy']);
-
-    Route::get('/users', [UserController::class, 'index']);
-    Route::post('/users', [UserController::class, 'store']);
-    Route::put('/users/{id}', [UserController::class, 'update']);
-    Route::delete('/users/{id}', [UserController::class, 'destroy']);
+Route::middleware(['auth:sanctum'])->group(function () {
+    Route::get('/company', [CompanyController::class, 'index'])->middleware('role:Administrator|Manager');
+    Route::post('/company', [CompanyController::class, 'store'])->middleware('role:Administrator');
+    Route::put('/company/{id}', [CompanyController::class, 'update'])->middleware('role:Administrator|Manager');
+    Route::delete('/company/{id}', [CompanyController::class, 'destroy'])->middleware('role:Administrator');
+});
